Правка замечаний (часть 1): добавлено больше методов, генерация рандомных данных в бенчмарках, сохранение вычисления хэша в кэше

+++ src/main/java/ru/spb/itmo/asashina/lab1/perf/hash/CreatingBenchmarkRunner.java
@@ -45,7 +45,7 @@ public class CreatingBenchmarkRunner {
             k++;
             tempValue /= Integer.MAX_VALUE;
         }
-        return (int) (value.hashCode() + k);
+        return value.hashCode() + k;
     }
 
 }
